% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrapper.R
\name{BLRM.fit.mwg}
\alias{BLRM.fit.mwg}
\title{Bayesian Logistic Regression Model (BLRM) training}
\usage{
BLRM.fit.mwg(
  Y0,
  X0,
  PriorVar,
  propSD0,
  nMC = 1000,
  nBI = 250,
  thin = 5,
  seed = 1
)
}
\arguments{
\item{Y0}{vector of responses}

\item{X0}{covariate matrix}

\item{PriorVar}{variance of prior distribution of beta}

\item{propSD0}{vector of standard deviations for normal proposal density}

\item{nMC}{number of MCMC samples}

\item{nBI}{number of burn-in samples}

\item{thin}{number of samples to skip over in thinning}

\item{seed}{set seed for random number generation}
}
\value{
a nested list of beta samples, and beta acceptance rates
}
\description{
Performs Bayesian Logistic Regression Model training by sampling beta
from posterior distribution with user specified parameters and data
}
\examples{

## simulate data;

set.seed(1);
N  = 100;
p  = 10;

X  = matrix(data = rnorm(N*p), nrow=N, ncol=p)
beta_true = c(rep(1,p/2),rep(0,p/2))
eta = X \%*\% beta_true
pi = exp(eta) / (1 + exp(eta))
Y  = rbinom(N,1,pi)
propSD = rep(1,p)

## fit model;
test1 <- G3proj::BLRM.fit.mwg(Y0 = Y, X0 = X, PriorVar = 1000, propSD0 = propSD,
                      nMC = 500, nBI = 100, thin = 5)
}
