% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrapper.R
\name{BLRM.Tuning}
\alias{BLRM.Tuning}
\title{Bayesian Logistic Regression Model (BLRM) Tuning for Proposal Standard Deviation}
\usage{
BLRM.Tuning(Y0, X0, PriorVar, nMC = 10000, b = 50, seed = 1)
}
\arguments{
\item{Y0}{vector of responses}

\item{X0}{covariate matrix}

\item{PriorVar}{variance of prior distribution of beta}

\item{nMC}{number of MCMC samples, should be the multiple of b}

\item{b}{batch size}

\item{seed}{set seed for random number generation}
}
\value{
a nested list of beta samples, beta acceptance rates and proposal SD for each batch
}
\description{
Performs hyper parameter tuning of proposal standard deviation
that can be used in Bayesian Logistic Regression Model training
with user specified parameters and data
}
\details{
Runs nMc/b number of batches of size b. In each batch the algorithm calculates
the Metropolis acceptance rate and adjusts the proposal standard deviation
accordingly. As such, nMC must be a multiple of b.
}
\examples{

## simulate data;

set.seed(1);
N  = 100;
p  = 10;

X  = matrix(data = rnorm(N*p), nrow=N, ncol=p)
beta_true = c(rep(1,p/2),rep(0,p/2))
eta = X \%*\% beta_true
pi = exp(eta) / (1 + exp(eta))
Y  = rbinom(N,1,pi)

## Compute optimal proposal SD;
test1 <- BLRM.Tuning(Y0 = Y, X0 = X, PriorVar = 1000,
                     nMC = 100, b = 10)
}
