% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso-function.R
\name{glmlasso}
\alias{glmlasso}
\title{Lasso regression for logistics model}
\usage{
glmlasso(X, y, lambda = 0.1, tol = 1e-06, iter = 100)
}
\arguments{
\item{X}{Design Matrix}

\item{y}{binary outcome}

\item{lambda}{Lasso penalty parameter}

\item{tol}{convergence threshold}

\item{iter}{iteration times}
}
\value{
Lasso estimate of coefficients
}
\description{
Lasso regression for logistics model
}
\examples{
set.seed(1232)
Nz = 500
pz = 10
Xz = scale(matrix(rnorm(Nz*pz), ncol=pz))
bz = c(.5, -.5, .25, -.25, .125, -.125, rep(0, pz-6))
yz = rbinom(Nz,1,exp(Xz \%*\% bz)/(1+exp(Xz \%*\% bz)))
lambda = .1
require(glmnet)
fit <- glmnet(Xz,yz,family="binomial",lambda = 0.1,intercept = FALSE)
coef(fit,s = 0.1)
fit1 <- glmlasso(Xz,yz,lambda,tol=1e-12)



}
