% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso-function.R
\name{optim.lambda}
\alias{optim.lambda}
\title{select optimal tuning parameter lambda for lasso based on BIC criterion.}
\usage{
optim.lambda(Xz, yz, lambda.min, lambda.max, len, plot = FALSE)
}
\arguments{
\item{Xz}{the design matrix for glmlasso}

\item{yz}{a vector of binary outcomes}

\item{lambda.min}{the smallest lambda}

\item{lambda.max}{the largest lambda}

\item{len}{the number of lambdas for fitting}

\item{plot}{logic variable, indicating if the lambda versus BIC plot should be returned}
}
\value{
a plot of BIC vs lambda and the optimal lambda with the smallest BIC
}
\description{
This function will fit the glmlasso with a sequence value of tuning parameter lambdas and
return a plot of BIC vs lambda and the optimal lambda with the smallest BIC.
}
\details{
This function will need glmlasso function.
}
\examples{

set.seed(8675309)
Xz = scale(matrix(rnorm(5000), ncol=10))
bz = c(.5, -.5, .25, -.25, .125, -.125, rep(0, 4))
yz = rbinom(Nz,1,exp(Xz \%*\% bz)/(1+exp(Xz \%*\% bz)))
optim.lambda(Xz,yz,lambda.min = 0,lambda.max = 0.1,len = 200)


}
