% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso-function.R
\name{predict_glmlasso}
\alias{predict_glmlasso}
\title{Prediction of Lasso logistics regression}
\usage{
predict_glmlasso(fit, newdata, type = "response", threshold = 0.5)
}
\arguments{
\item{fit}{fitted object from glmlasso}

\item{newdata}{a data frame in which to look for variables with which to predict.}

\item{type}{the type of prediction required. The default is on the scale of the response variable.
\code{type="class"} gives the predicted classification result}

\item{threshold}{Threshold values decided for classification}
}
\value{
if \code{type="class"} a vector of predicted class will be returned. Otherwise return a vector of raw probability.
}
\description{
Prediction of Lasso logistics regression
}
\examples{
set.seed(1232)
Nz = 500
pz = 10
Xz = scale(matrix(rnorm(Nz*pz), ncol=pz))
bz = c(.5, -.5, .25, -.25, .125, -.125, rep(0, pz-6))
yz = rbinom(Nz,1,exp(Xz \%*\% bz)/(1+exp(Xz \%*\% bz)))
lambda = .1
fit1 <- glmlasso(Xz,yz,lambda,tol=1e-12)
pred <- predict_glmlasso(fit1,newdata = Xz)

}
