% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso-function.R
\name{soft_thresh}
\alias{soft_thresh}
\title{Soft thresholding calculation}
\usage{
soft_thresh(x, b)
}
\arguments{
\item{x}{a numeric vector}

\item{b}{threshold values}
}
\value{
a numeric vector with the same length as x after soft thresholding
}
\description{
Soft thresholding calculation
}
\examples{
set.seed(1)
x <- rnorm(5)
b <- 0.2
soft_thresh(x,b)



}
