% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrapper.R
\name{SSVS.Logistic}
\alias{SSVS.Logistic}
\title{Stochastic Search Variable Selection (SSVS) for Logistic Regression}
\usage{
SSVS.Logistic(
  Y0,
  X0,
  propSD0,
  c0,
  tau0,
  nMC = 1000L,
  nBI = 250L,
  thin = 5L,
  seed = 1L
)
}
\arguments{
\item{Y0}{vector of responses}

\item{X0}{covariate matrix without intercept}

\item{propSD0}{vector of standard deviations for normal proposal density}

\item{c0}{parameter for spike and slab prior of beta}

\item{tau0}{parameter for spike and slab prior of beta}

\item{nMC}{number of MCMC samples}

\item{nBI}{number of burn-in samples}

\item{thin}{number of samples to skip over in thinning}

\item{seed}{set seed for random number generation}
}
\value{
a nested list of gamma samples, beta samples, and beta acceptance rates
}
\description{
Performs SSVS for a logistic regression model with user specified parameters and data
}
\examples{

## simulate data

set.seed(1)
N  = 100
p  = 10

X  = matrix(data = rnorm(N*p), nrow=N, ncol=p)
beta_true = c(rep(1,p/2),rep(0,p/2))
eta = X \%*\% beta_true
pi = exp(eta) / (1 + exp(eta))
Y  = rbinom(N,1,pi)
propSD0 = rep(.5,p)

## fit model;
test1 <- G3proj::SSVS.Logistic(Y0 = Y, X0 = X, propSD0, c0 = 10,
                               tau0 = 0.4, nMC = 1000, nBI = 100, thin=1, seed=1)
}
