% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrapper.R
\name{SSVS.Tuning}
\alias{SSVS.Tuning}
\title{MH Random Walk Proposal Tuning for SSVS under Logistic Regression Model}
\usage{
SSVS.Tuning(Y0, X0, c0, tau0, nMC = 1000, b = 50, seed = 1)
}
\arguments{
\item{Y0}{vector of responses}

\item{X0}{covariate matrix without intercept}

\item{c0}{parameter for spike and slab prior of beta}

\item{tau0}{parameter for spike and slab prior of beta}

\item{nMC}{number of MCMC samples}

\item{b}{batch size}

\item{seed}{set seed for random number generation}
}
\value{
A nested list of gamma samples, beta samples, beta acceptance rates for each batch, and the proposal standard deviations at the final batch.
}
\description{
Tunes the normal proposal distribution standard deviation for Metropolis Hastings MCMC for the SSVS variable selection method in a Logistic Regression model.
}
\details{
Runs nMc/b number of batches of size b. In each batch the algorithm calculates the Metropolis acceptance rate and adjusts the proposal standard deviation accordingly. As such, nMC must be a multiple of b.
}
\examples{

## simulate data;

set.seed(1);
N  = 100;
p  = 10;

X  = matrix(data = rnorm(N*p), nrow=N, ncol=p)
beta_true = c(rep(1,p/2),rep(0,p/2))
eta = X \%*\% beta_true
pi = exp(eta) / (1 + exp(eta))
Y  = rbinom(N,1,pi)

## fit model;
test1 <- G3proj::SSVS.Tuning(Y0 = Y, X0 = X, c0 = 10,
                             tau0 = 0.4, nMC = 1000, b = 50)
}
